Adding MapStruct, Moving Mapping logic to Service Level
