+++ src/main/java/com/csmarton/elasticsearch/mapper/UserMapper.java
@@ -1,0 +1,12 @@
+package com.csmarton.elasticsearch.mapper;
+
+import com.csmarton.elasticsearch.model.CreateUserRequest;
+import com.csmarton.elasticsearch.model.User;
+import org.mapstruct.Mapper;
+
+@Mapper(componentModel = "spring")
+public interface UserMapper
+{
+
+    User mapUserFromRequest(CreateUserRequest request);
+}
