+++ src/main/java/com/csmarton/elasticsearch/service/UserService.java
@@ -2,12 +2,13 @@ package com.csmarton.elasticsearch.service;
 
 import com.csmarton.elasticsearch.model.CreateUserRequest;
 import com.csmarton.elasticsearch.model.User;
+import com.csmarton.elasticsearch.model.UserSearchResponse;
 
 import java.util.List;
 
 public interface UserService {
     User createUser(CreateUserRequest userRequest);
 
-    List<User> searchUser(String firstName, String lastname);
+    UserSearchResponse searchUser(String firstName, String lastname);
 
 }
