+++ src/main/java/com/csmarton/elasticsearch/service/UserServiceImpl.java
@@ -1,7 +1,10 @@
 package com.csmarton.elasticsearch.service;
 
+import com.csmarton.elasticsearch.mapper.UserMapper;
 import com.csmarton.elasticsearch.model.CreateUserRequest;
 import com.csmarton.elasticsearch.model.User;
+import com.csmarton.elasticsearch.model.UserResponse;
+import com.csmarton.elasticsearch.model.UserSearchResponse;
 import com.csmarton.elasticsearch.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -10,6 +13,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -18,16 +22,14 @@ public class UserServiceImpl implements UserService {
 
     private final UserRepository userRepository;
 
+    private final UserMapper userMapper;
+
     @Override
     public User createUser(CreateUserRequest userRequest) {
 
         log.debug("Creating user - Begin");
 
-        User user = User.builder()
-                .firstName(userRequest.getFirstName())
-                .lastName(userRequest.getLastName())
-                .email(userRequest.getEmail())
-                .build();
+        User user = this.userMapper.mapUserFromRequest(userRequest);
 
         User result = userRepository.save(user);
         log.debug("User created with uuid: {}", result.getUuid());
@@ -38,7 +40,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public List<User> searchUser(String firstName, String lastname) {
+    public UserSearchResponse searchUser(String firstName, String lastname) {
         log.debug("Searching User - Begin");
 
         List<User> users = null;
@@ -53,8 +55,23 @@ public class UserServiceImpl implements UserService {
 
         log.debug("Result of user search: Found {} record(s)", users.size());
 
+        UserSearchResponse userSearchResponse = createUserSearchResponse(users);
+
         log.debug("Searching User - End");
 
-        return users;
+        return userSearchResponse;
+    }
+
+    private UserSearchResponse createUserSearchResponse(List<User> users) {
+        UserSearchResponse userSearchResponse = new UserSearchResponse();
+
+        userSearchResponse.setUsers(
+                users.stream()
+                        .map(user -> new UserResponse(user.getUuid(), user.getFirstName(), user.getLastName(), user.getEmail() ))
+                        .collect(Collectors.toList()));
+
+        userSearchResponse.setTotal(users.size());
+
+        return userSearchResponse;
     }
 }
